+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 3.4.2b, 2015-05-13
 ------------------
+- added emoticon statistics to HTML export;
 - applying or resetting message filter will scroll to last selection;
 - added support for copying selected list items to clipboard;
 - avoid caching messages on export to decrease memory overhead;